+++ asyncz/schedulers/__init__.py
@@ -1,3 +1,4 @@
 from .asyncio import AsyncIOScheduler, NativeAsyncIOScheduler
+from .inspection import SchedulerInfo
 
-__all__ = ["AsyncIOScheduler", "NativeAsyncIOScheduler"]
+__all__ = ["AsyncIOScheduler", "NativeAsyncIOScheduler", "SchedulerInfo"]